[SmartRobot]should_save_in_1st_locker_when_locker1_has_1_locker2_has_2

+++ src/test/java/PrimaryRobotTest.java
@@ -37,7 +37,7 @@ public class PrimaryRobotTest {
 
         //Then:
         assertEquals("001", ticket.getBoxId());
-        assertEquals(1L, locker1.getAvailableBox());
+        assertEquals(1L, (long)locker1.getAvailableBox());
     }
 
 
@@ -59,7 +59,7 @@ public class PrimaryRobotTest {
 
         //Then:
         assertEquals("003", ticket.getBoxId());
-        assertEquals(1L, locker2.getAvailableBox());
+        assertEquals(1L, (long)locker2.getAvailableBox());
     }
 
     @Test
